change tipes to common

+++ migrations/20221103103601_create_movie_table.sql
@@ -3,8 +3,8 @@ CREATE TABLE movies(
    id INTEGER PRIMARY KEY NOT NULL,
    title TEXT NOT NULL,
    director TEXT NOT NULL,
-   relise_date SMALLINT NOT NULL,
-   rating TINYINT NOT NULL,
+   relise_date INTEGER NOT NULL,
+   rating INTEGER NOT NULL,
    poster_id TEXT NOT NULL,
    created_at DATETIME NOT NULL
 );
